fix
fix: resolvido acesso a algumas rotas

+++ app/services/SongService.ts
@@ -20,12 +20,12 @@ export const getAllSongs = async () => {
 };
 
 export const updateSong = async (id: number, song: any) => {
-    const response = await api.put(`v1/songs/${id}`, song);
+    const response = await api.put(`v1/admin/songs/${id}`, song);
     return response.data;
 };
 
 export const deleteSong = async (id: number) => {
-    const response = await api.delete(`v1/songs/${id}`);
+    const response = await api.delete(`v1/admin/songs/${id}`);
     return response.data;
 };
